添加用户和vip管理

+++ src/main/java/cn/guba/igu8/db/mysqlModel/base/BaseTeacher.java
@@ -54,12 +54,12 @@ public abstract class BaseTeacher<M extends BaseTeacher<M>> extends Model<M> imp
 		return get("pfVipUid");
 	}
 
-	public M setVipTypeId(java.lang.Long vipTypeId) {
+	public M setVipTypeId(java.lang.Integer vipTypeId) {
 		set("vipTypeId", vipTypeId);
 		return (M)this;
 	}
 
-	public java.lang.Long getVipTypeId() {
+	public java.lang.Integer getVipTypeId() {
 		return get("vipTypeId");
 	}
