+++ src/main/java/cn/guba/igu8/db/mysqlModel/base/BaseViptype.java
@@ -9,12 +9,12 @@ import com.jfinal.plugin.activerecord.IBean;
 @SuppressWarnings({"serial", "unchecked"})
 public abstract class BaseViptype<M extends BaseViptype<M>> extends Model<M> implements IBean {
 
-	public M setId(java.lang.Long id) {
+	public M setId(java.lang.Integer id) {
 		set("id", id);
 		return (M)this;
 	}
 
-	public java.lang.Long getId() {
+	public java.lang.Integer getId() {
 		return get("id");
 	}
